feat: launch url changed; impliment city controler and action status; mapper injected

+++ PeoplesControl/Logic/Repositories/Interfaces/IActionMetaRepository.cs
@@ -7,7 +7,7 @@ namespace Logic.Repositories
 {
     public interface IActionMetaRepository
     {
-        public ActionMeta Get(int id);
+        public ActionMeta Get(long id);
         public List<ActionMeta> GetAll();
         public ActionMeta Add(ActionMeta actionMeta);
     }
